Fixed TitleContainer to use partial classes for platform support.
