+++ MonoGame.Framework/TitleContainer.Android.cs
@@ -1,0 +1,18 @@
+// MonoGame - Copyright (C) The MonoGame Team
+// This file is subject to the terms and conditions defined in
+// file 'LICENSE.txt', which is part of this source code package.
+
+using System;
+using System.IO;
+
+namespace Microsoft.Xna.Framework
+{
+    partial class TitleContainer
+    {
+        private static Stream PlatformOpenStream(string name, string safeName)
+        {
+            return Android.App.Application.Context.Assets.Open(safeName);
+        }
+    }
+}
+
