+++ MonoGame.Framework/TitleContainer.Desktop.cs
@@ -1,0 +1,26 @@
+// MonoGame - Copyright (C) The MonoGame Team
+// This file is subject to the terms and conditions defined in
+// file 'LICENSE.txt', which is part of this source code package.
+
+using System;
+using System.IO;
+
+namespace Microsoft.Xna.Framework
+{
+    partial class TitleContainer
+    {
+        static partial void PlatformInit()
+        {
+#if WINDOWS || DESKTOPGL
+            Location = AppDomain.CurrentDomain.BaseDirectory;
+#endif
+        }
+
+        private static Stream PlatformOpenStream(string name, string safeName)
+        {
+            var absolutePath = Path.Combine(Location, safeName);
+            return File.OpenRead(absolutePath);
+        }
+    }
+}
+
